requirements.txt
qdrant-client
sentence-transformers
openai
gradio
langgraph
langchain-core
langchain-anthropic
mcp
pyyaml
requests
sympy
pillow
anthropic
